Implement GetStat() function.
This implementation covers support for both files and directories.

In order to implement GetStat() support for directories, the internal
function CDVD_findfile() has been slighty modified to return a
non-zero value when a directory with the specified path is found.

Note that its UDF counterpart - UDF_FindFile() - does not need any
modifications, as it returns non-zero when either a file or a
directory with the specified path is found.